+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.4.2 (July 13, 2022)
+* Enhancements
+  * Updated argparse decorator to remove annotations when the docstring is used for a command's help text.
+  * Updated unit test to be Python 3.11 compliant.
+
+## 2.4.1 (April 13, 2022)
+* Bug Fixes
+  * Fixed value for `ansi.Bg.YELLOW`.
+  * Fixed unit tests for `ansi.allow_style`.
+* Enhancements
+  * `async_alert()` raises a `RuntimeError` if called from the main thread.
+
 ## 2.4.0 (February 22, 2022)
 * Bug Fixes
   * Fixed issue in `ansi.async_alert_str()` which would raise `IndexError` if prompt was blank.